+++ src/cpp/libberadio/simulavr.h
@@ -1,0 +1,7 @@
+// SimulAVR debug statement shortcuts
+// http://www.nongnu.org/simulavr/intro.html
+void _l(std::string str);
+void _d(int val);
+void _d(double val);
+void _d(const char *str);
+void _d(std::string str);