+++ CHANGES.rst
@@ -13,6 +13,7 @@ develop
 - add BERadioMessage as a convenient message builder, with api docs and doctests
 - add test/comparison.rst as a shootout for comparing space efficiency of different
   serialization mechanisms (BERadio vs. Bencode vs. Binary vs. CSV vs. JSON vs. YAML)
+- add complete c++ spike using variadic arguments and stl vectors, also runs in simulavr
 
 
 2015-10-27 0.4.4